requirements.txt
Flask==2.1.2
Flask-CORS==3.0.10
tensorflow==2.8.0
tensorflow-hub==0.12.0
numpy==1.21.5
Pillow==9.0.1
